Fixed issue #72
