+++ src/main/java/fi/otavanopisto/kuntaapi/server/controllers/EventController.java
@@ -3,6 +3,7 @@ package fi.otavanopisto.kuntaapi.server.controllers;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
 
@@ -15,6 +16,8 @@ import fi.otavanopisto.kuntaapi.server.id.EventId;
 import fi.otavanopisto.kuntaapi.server.id.OrganizationId;
 import fi.otavanopisto.kuntaapi.server.integrations.AttachmentData;
 import fi.otavanopisto.kuntaapi.server.integrations.EventProvider;
+import fi.otavanopisto.kuntaapi.server.integrations.EventProvider.EventOrder;
+import fi.otavanopisto.kuntaapi.server.integrations.EventProvider.EventOrderDirection;
 import fi.otavanopisto.kuntaapi.server.rest.model.Attachment;
 import fi.otavanopisto.kuntaapi.server.rest.model.Event;
 
@@ -70,9 +73,20 @@ public class EventController {
       Integer firstResult, Integer maxResults, EventProvider.EventOrder order,
       EventProvider.EventOrderDirection orderDirection, OrganizationId organizationId) {
     List<Event> result = new ArrayList<>();
+    
     for (EventProvider eventProvider : getEventProviders()) {
-      result.addAll(eventProvider.listOrganizationEvents(organizationId, startBefore, startAfter, endBefore, endAfter, order, orderDirection, firstResult, maxResults));
+      result.addAll(eventProvider.listOrganizationEvents(organizationId, startBefore, startAfter, endBefore, endAfter));
     }
+    
+    int resultCount = result.size();
+    int firstIndex = firstResult == null ? 0 : Math.min(firstResult.intValue(), resultCount);
+    int toIndex = maxResults == null ? resultCount : Math.min(firstIndex + maxResults.intValue(), resultCount);
+    
+    return sortEvents(result.subList(firstIndex, toIndex), order, orderDirection);
+  }
+  
+  private List<Event> sortEvents(List<Event> result, EventProvider.EventOrder order, EventProvider.EventOrderDirection orderDirection) {
+    Collections.sort(result, new EventComparator(order, orderDirection));
     return result;
   }
   
@@ -86,4 +100,61 @@ public class EventController {
     
     return Collections.unmodifiableList(result);
   } 
+  
+  private class EventComparator implements Comparator<Event> {
+    
+    private EventOrder order;
+    private EventOrderDirection direction;
+    
+    public EventComparator(EventOrder order, EventOrderDirection direction) {
+      this.order = order;
+      this.direction = direction;
+    }
+    
+    @Override
+    public int compare(Event event1, Event event2) {
+      int result;
+      
+      switch (order) {
+        case END_DATE:
+          result = compareEndDates(event1, event2);
+        break;
+        case START_DATE:
+          result = compareStartDates(event1, event2);
+        break;
+        default:
+          result = 0;
+        break;
+      }
+      
+      if (direction == EventOrderDirection.ASCENDING) {
+        return -result;
+      } 
+      
+      return result;
+    }
+
+    private int compareStartDates(Event event1, Event event2) {
+      return compareDates(event1.getStart(), event2.getStart());
+    }
+
+    private int compareEndDates(Event event1, Event event2) {
+      return compareDates(event1.getEnd(), event2.getEnd());
+    }
+    
+    private int compareDates(OffsetDateTime dateTime1, OffsetDateTime dateTime2) {
+      if (dateTime1 == null && dateTime2 == null) {
+        return 0;
+      }
+      
+      if (dateTime1 == null) {
+        return 1;
+      } else if (dateTime2 == null) {
+        return -1;
+      }
+              
+      return dateTime1.compareTo(dateTime2);
+    }
+    
+  }
 }
