+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/EventProvider.java
@@ -24,15 +24,10 @@ public interface EventProvider {
    * @param startAfter return only events starting after the date
    * @param endBefore return only events ending before the date
    * @param endAfter return only events ending after the date
-   * @param order order
-   * @param orderDirection ordering direction
-   * @param firstResult first index of results
-   * @param maxResults maximum number of results
    * @return events organization events
    */
   public List<Event> listOrganizationEvents(OrganizationId organizationId, OffsetDateTime startBefore, OffsetDateTime startAfter, 
-      OffsetDateTime endBefore, OffsetDateTime endAfter, EventOrder order, EventOrderDirection orderDirection, 
-      Integer firstResult, Integer maxResults);
+      OffsetDateTime endBefore, OffsetDateTime endAfter);
   
   /**
    * Finds a single organization event
